Add functionality to button 1

+++ 11GGithub/View/MainView.cs
@@ -19,9 +19,9 @@ namespace _11GGithub.View
 
         private void btn1_Click(object sender, EventArgs e)
         {
-            int a = Int32.Parse(txtA.Text);
-            int b = Int32.Parse(txtB.Text);
-            int output = a / b;
+            double a = Int32.Parse(txtA.Text);
+            double b = Int32.Parse(txtB.Text);
+            double output = a / b;
             lblOutput.Text = (" " + output);
         }
